+++ react-training/src/app/components/App/index.js
@@ -5,15 +5,13 @@ import Account from '../../screens/Account';
 import AuthenticatedRoute from '../AuthenticatedRoute';
 import Game from '../../screens/Game';
 import Login from '../../screens/Login';
-import Logout from '../Logout';
-import { routes } from '../../constants/routes';
+import { routes } from '../../../constants/routes';
 
 const App = () => (
   <Router>
     <Switch>
       <AuthenticatedRoute path={routes.login} component={Login} />
       <AuthenticatedRoute path={routes.account} component={Account} />
-      <AuthenticatedRoute path={routes.logout} component={Logout} />
       {routes.game.map(route => (
         <AuthenticatedRoute key={route} path={route} component={Game} />
       ))}
