+++ react-training/src/app/components/AuthenticatedRoute/index.js
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
 import { Route, Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
 
 import authActions from '../../../redux/auth/action';
 import Login from '../../screens/Login';
-import { UNAUTHENTICATED_DEFAULT_ROUTE, AUTHENTICATED_DEFAULT_ROUTE } from '../../constants/auth';
+import { UNAUTHENTICATED_DEFAULT_ROUTE, AUTHENTICATED_DEFAULT_ROUTE } from '../../../constants/auth';
 import PrivatePage from '../PrivatePage';
 
 class AuthenticatedRoute extends Component {
@@ -39,12 +40,19 @@ const mapDispatchToProps = dispatch => ({
 
 const mapStateToProps = ({ auth }) => ({
   isAuth: auth.isAuth,
-  user: auth.user,
-  currentLocation: auth.currentLocation,
-  previousLocation: auth.previousLocation
+  user: auth.user
 });
 
 export default connect(
   mapStateToProps,
   mapDispatchToProps
 )(AuthenticatedRoute);
+
+AuthenticatedRoute.propTypes = {
+  component: PropTypes.func,
+  isAuth: PropTypes.bool,
+  location: PropTypes.shape({ location: PropTypes.string }),
+  path: PropTypes.string,
+  setAuthentication: PropTypes.func,
+  user: PropTypes.string
+};
