+++ react-training/src/app/components/CustomLink/index.js
@@ -1,10 +1,13 @@
 import { withRouter } from 'react-router-dom';
 import React, { Component } from 'react';
-import { Redirect } from 'react-router-dom';
 
 class CustomLink extends Component {
 
-  handleClick = (e) => this.props.history.push(this.props.route)
+  handleClick = (e) => {
+    const { route, handleCustomClick } = this.props;
+    if (handleCustomClick && typeof handleCustomClick === 'function') handleCustomClick();
+    else this.props.history.push(route);
+  };
 
   render() {
     const { tag } = this.props;
