+++ react-training/src/app/components/PrivatePage/index.js
@@ -1,11 +1,12 @@
 import React, { Fragment } from 'react';
-import Topbar from '../../screens/Topbar';
+
+import Topbar from '../Topbar';
 
 const PrivatePage = ({ component: Component, ...props }) => (
   <Fragment>
     <Topbar />
     <Component {...props} />
   </Fragment>
-)
+);
 
 export default PrivatePage;
