+++ react-training/src/app/components/Topbar/Menu/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../../scss/variables/colors.scss';
+@import 'variables/colors';
 
 .list {
   display: flex;
