+++ react-training/src/app/components/Topbar/index.js
