+++ react-training/src/app/components/Topbar/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/variables/colors.scss';
+@import 'variables/colors';
 
 .navbar {
   height: 90px;
