+++ react-training/src/app/screens/Account/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/variables/colors.scss';
+@import 'variables/colors';
 
 .main {
   align-items: center;
