+++ react-training/src/app/screens/Game/Square/styles.scss
@@ -1,4 +1,4 @@
-@import './variables/colors';
+@import 'variables/colors';
 
 .square {
   background: $white;
