+++ react-training/src/app/screens/Login/LoginForm/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../../scss/variables/colors.scss';
+@import 'variables/colors';
 
 .login {
   display: grid;
