+++ react-training/src/constants/auth.js
