+++ react-training/src/constants/lines.js
