+++ react-training/src/constants/routes.js
@@ -1,6 +1,5 @@
 export const routes = {
   login: '/login',
   game: ['/', '/game'],
-  account: '/account',
-  logout: '/logout'
+  account: '/account'
 };
