+++ react-training/src/scss/index.scss
@@ -1,6 +1,6 @@
 @import url('https://fonts.googleapis.com/css?family=Roboto:400,400i');
 @import './equalizer.scss';
-@import './variables/colors.scss';
+@import 'variables/colors';
 
 body {
   font: 14px 'Roboto', sans-serif;
